最新版に更新

+++ apps/frontend/my-app/src/Component/Contentinfo.js
@@ -1,23 +1,39 @@
-import React, { useContext } from 'react';
+import React, { useContext, useState } from 'react';
 import './Contentinfo.css'; // CSSファイルをインポート
 import Userinfo from './Userinfo';
 import Content from './Content';
 import Fav from './Fav';
 import deletePost from './DeletePost';
 import AuthContext from '../Utils/AuthProvider';
+import ReportPost from './ReportPost';
 
 function Contentinfo({ src, alt, username, userid, postid, content, imagepath, handleDelete }) {
   const { user } = useContext(AuthContext);
   const currentUserId = user?.id;
+  const [isReportOpen, setIsReportOpen] = useState(false);
 
-  const handleClick = async () => {
+  const handleDeleteClick = async () => {
+    const confirmDelete = window.confirm("本当にこのポストを削除しますか？");
+    if (!confirmDelete) {
+      return; // キャンセルされた場合は何もしない
+    }
+
+
     try {
       await deletePost(postid, userid);
       handleDelete(postid);
       alert('ポストが削除されました');
     } catch (error) {
       alert('ポストの削除に失敗しました');
-    }
+    } 
+  };
+
+  const handleReportClick = () => {
+    setIsReportOpen(true);
+  };
+
+  const handleReportClose = () => {
+    setIsReportOpen(false);
   };
 
   return (
@@ -30,10 +46,14 @@ function Contentinfo({ src, alt, username, userid, postid, content, imagepath, h
       </div>
       <div className="post-actions">
         <div className="Fav">
-        <Fav postid={postid} />
+          <Fav postid={postid} />
         </div>
+        <button className="post-report-button" onClick={handleReportClick}>
+          通報
+        </button>
+        {isReportOpen && <ReportPost postid={postid} onClose={handleReportClose} />}
         {currentUserId === userid && (
-          <button className="post-delete-button" onClick={handleClick}>
+          <button className="post-delete-button" onClick={handleDeleteClick}>
             削除
           </button>
         )}
